Refactor create-character page to use form component for reusability
